Expect cores to be a regular integer
[#120615661]

+++ system/system_test.go
@@ -244,7 +244,7 @@ power management:
 		mockFS.EXPECT().Read("/proc/cpuinfo").Return([]byte(cpuinfo), nil)
 		cores, err := sys.PhysicalCores()
 		Expect(err).NotTo(HaveOccurred())
-		Expect(cores).To(Equal(uint64(4)))
+		Expect(cores).To(Equal(4))
 	})
 
 	Context("when it cannot read cpuinfo", func() {
